Контролер AccountController розбито на три різні контролери для дотримання SRP
