+++ NotesApp/Controllers/ProfileController.cs
@@ -1,21 +1,46 @@
-using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
-using System.Threading.Tasks;
+using NotesApp.Models;
 
-namespace NotesApp.Controllers
+[Authorize]
+[Route("Account")]
+public class ProfileController : Controller
 {
-    [Authorize]
-    public class ProfileController : BaseController
-    {
-        public ProfileController(UserManager<IdentityUser> userManager) : base(userManager)
-        {
-        }
+    private readonly UserManager<IdentityUser> _userManager;
 
-        public async Task<IActionResult> Index()
+    public ProfileController(UserManager<IdentityUser> userManager)
+    {
+        _userManager = userManager;
+    }
+
+    private async Task<IdentityUser> GetCurrentUserAsync()
+      => await _userManager.GetUserAsync(User);
+
+    [HttpGet("Profile")]
+    public async Task<IActionResult> Profile()
+    {
+        var user = await GetCurrentUserAsync();
+        if (user == null) return NotFound();
+
+        return View(new ProfileViewModel
         {
-            var user = await GetCurrentUserAsync();
-            return user == null ? NotFound() : View(user);
-        }
+            Username = user.UserName,
+            Email = user.Email,
+            EmailConfirmed = user.EmailConfirmed
+        });
+    }
+
+    [HttpPost("VerifyEmailPost")]
+    [ValidateAntiForgeryToken]
+    public async Task<IActionResult> VerifyEmailPost()
+    {
+        var user = await GetCurrentUserAsync();
+        if (user == null) return NotFound();
+
+        user.EmailConfirmed = true;
+        await _userManager.UpdateAsync(user);
+
+        return RedirectToAction("Profile");
     }
 }
